qwen-vl-finetune/requirements.txt
accelerate==1.13.0
certifi==2026.2.25
charset-normalizer==3.4.6
einops==0.8.2
filelock==3.18.0
# flash_attn @ file://soft/flash_attn-2.8.1%2Bcu12torch2.7cxx11abiFALSE-cp310-cp310-linux_x86_64.whl#sha256=2b607ac503f5913f4aecd62408d0745ad5af6ecc59d4f96f90c30d3fe7a9e9dd
fsspec==2025.5.1
hf-xet==1.4.2
huggingface_hub==0.36.2
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==26.0
pillow==11.3.0
psutil==7.2.2
PyYAML==6.0.3
regex==2026.3.32
requests==2.33.1
safetensors==0.7.0
scipy==1.15.3
sympy==1.14.0
tokenizers==0.22.2
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.3
transformers==4.57.0
triton==3.3.0
typing_extensions==4.14.0
urllib3==2.6.3
